Add missing stations, improve handling of missing files/entries
Added a few stations which were missing. Instead of skipping the entire line for entries which had malformed or missing data, now I replace these with 0. If a weather station file is missing for an older date (i.e. we cannot download it and we don't have a saved version) now it will only raise a warning rather than crashing.